+++ src/Mapbender/VectorTilesBundle/Component/VectorTilesInstanceFactory.php
@@ -59,9 +59,12 @@ class VectorTilesInstanceFactory extends SourceInstanceFactory
         $instance->setMaxScale($data['maxScale'] ?? null);
         $instance->setFeatureInfo($data['featureInfo'] ?? true);
         $instance->setFeatureInfoAllowToggle($data['featureInfoAllowToggle'] ?? true);
-        $instance->setPropertyMap(($data['propertyMap'] ?? null) ? json_encode($data['propertyMap']) : null);
+        $instance->setFeatureInfoPropertyMap(($data['featureInfoPropertyMap'] ?? null) ? json_encode($data['featureInfoPropertyMap']) : null);
         $instance->setHideIfNoTitle($data['hideIfNoTitle'] ?? true);
         $instance->setFeatureInfoTitle($data['featureInfoTitle'] ?? null);
+        $instance->setPrintScaleCorrection($data['printScaleCorrection'] ?? 1.0);
+        $instance->setLegendEnabled($data['legend'] ?? $data['legendEnabled'] ?? false);
+        $instance->setLegendPropertyMap(($data['legendPropertyMap'] ?? null) ? json_encode($data['legendPropertyMap']) : null);
         if (isset($data['bbox']) && is_array($data['bbox'])) {
             $source->setBbox($data['bbox']);
         }